dashboard/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
greenlet==3.0.3
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.0.2
jsmin==3.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
lesscpy==0.15.1
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
oauthlib==3.2.2
ply==3.11
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.1.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
SQLAlchemy==1.4.29
typing_extensions==4.9.0
urllib3==2.2.0
websocket-client==1.7.0
Werkzeug==2.0.3
WTForms==3.0.0
xxhash==3.4.1
zipp==3.17.0